+++ Problem List.md
@@ -47,7 +47,7 @@
 | readiness_probe_misconfiguration_social_network     | Misconfiguration Failure            | New      | Pods in `user-service` will never enter `ready` state, making service unable to deal with requests                                                                                                                                | Cluster Management | App           |
 | resource_request_too_large                          | Cloud Management System Failure     | New      | Pods in `mongodb-rare` will never be scheduled because the container requires memory that excess limit in every node                                                                                                              | Container          | App           |
 | resource_request_too_small                          | Cloud Management System Failure     | New      | Pods in `mongodb-rare` will never be scheduled because the container requires memory that excess limit in every node                                                                                                              | Container          | App           |
-| hpa_missing_effective_cpu_request_hotel_reservation | Misconfiguration Failure | New | The `frontend` HorizontalPodAutoscaler targets CPU utilization but cannot compute the metric because the target pods lack an effective CPU request, so it reports `<unknown>/60%`, `ScalingActive=False`, and `FailedGetResourceMetric` while pods may remain Running/Ready. | Cluster Management | App |
+| hpa_missing_effective_cpu_request_hotel_reservation | Misconfiguration Failure            | New      | The `frontend` HPA cannot compute CPU utilization because target pods lack an effective CPU request, causing `ScalingActive=False` while pods remain Running/Ready                                                                 | Cluster Management | App           |
 | rolling_update_misconfigured_hotel_reservation      | Misconfiguration Failure            | New      | No pods will be available since they get stuck with a contradictory update configuration                                                                                                                                         | Cluster Management | App           |
 | rolling_update_misconfigured_social_network         | Misconfiguration Failure            | New      | No pods will be available since they get stuck with a contradictory update configuration                                                                                                                                         | Container          | App           |
 | scale_pod_zero_social_net                           | Cloud Management System Failure     | AIOpsLab | The replicas will be set to `0`, leading to no available pods, thus making service unavailable                                                                                                                                    | Cluster Management | App           |
